Audio Transcript ranslation with Whisper/requirements.txt
openai==0.28
python-dotenv
flask
openai-whisper
